+++ src/backend/joseki.be/webapp/GlobalSuppressions.cs
@@ -11,4 +11,4 @@
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1401:Fields should be private", Justification = "umut", Scope = "member", Target = "~F:webapp.Controllers.Data.Components")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1401:Fields should be private", Justification = "umut", Scope = "member", Target = "~F:webapp.Controllers.Data.OverallCounters")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1401:Fields should be private", Justification = "umut", Scope = "member", Target = "~F:webapp.Controllers.Data.ComponentSummaries")]
-[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1401:Fields should be private", Justification = "<Pending>", Scope = "member", Target = "~F:webapp.Controllers.Data.GetCounters")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1401:Fields should be private", Justification = "umut", Scope = "member", Target = "~F:webapp.Controllers.Data.GetCounters")]
