+++ src/backend/joseki.be/webapp/webapp.xml
@@ -28,7 +28,10 @@
         </member>
         <member name="F:webapp.Controllers.Data.Dates">
             <summary>
-            list of consistent dates for ui tests.
+            list of dates for ui tests.
+            date behaves like a key while comparing scans.
+            thus an overall scan and a component scan
+            must all have the same initial scan date associated.
             </summary>
         </member>
         <member name="F:webapp.Controllers.Data.Components">
@@ -36,12 +39,12 @@
             list of inftastructure components.
             </summary>
         </member>
-        <member name="F:webapp.Controllers.Data.Counters">
+        <member name="F:webapp.Controllers.Data.GetCounters">
             <summary>
             list of counter summaries.
             </summary>
         </member>
-        <member name="M:webapp.Controllers.Data.ScoreHistory(webapp.Models.InfrastructureComponent)">
+        <member name="M:webapp.Controllers.Data.GetScoreHistory(webapp.Models.InfrastructureComponent)">
             <summary>
             score history, combination of dates and scores.
             </summary>
@@ -53,9 +56,7 @@
         </member>
         <member name="M:webapp.Controllers.Data.GetComponentSummaries">
             <summary>
-            list of components (clusters and subscriptions).
-            this list is long because it's mocking data change over time.
-            in a real time scenario, only data on related date should be returned.
+            Returns summary for all components except overall.
             </summary>
         </member>
         <member name="F:webapp.Controllers.Data.overallComponent">
@@ -88,6 +89,26 @@
             <param name="start">Ticks on a start.</param>
             <param name="stop">Ticks in the end.</param>
             <returns>Elapsed milliseconds.</returns>
+        </member>
+        <member name="T:webapp.Models.ScanResult">
+            <summary>
+            Scan result enum.
+            </summary>
+        </member>
+        <member name="F:webapp.Models.ScanResult.NotFound">
+            <summary>
+            Enum value when a scan was not found
+            </summary>
+        </member>
+        <member name="F:webapp.Models.ScanResult.Failed">
+            <summary>
+            Enum value when a scan failed
+            </summary>
+        </member>
+        <member name="F:webapp.Models.ScanResult.Succeeded">
+            <summary>
+            Enum value when a scan succeeded
+            </summary>
         </member>
         <member name="T:webapp.Models.CountersSummary">
             <summary>
@@ -177,11 +198,6 @@
             Summarizes a single infrastructure unit stats: score, trends, name.
             </summary>
         </member>
-        <member name="P:webapp.Models.InfrastructureComponentSummary.Id">
-            <summary>
-            The unique Id of a component summary.
-            </summary>
-        </member>
         <member name="P:webapp.Models.InfrastructureComponentSummary.Date">
             <summary>
             The generation date of the Summary.
@@ -197,21 +213,60 @@
             Latest known check-result counters.
             </summary>
         </member>
-        <member name="P:webapp.Models.InfrastructureComponentSummary.ScoreHistory">
+        <member name="T:webapp.Models.InfrastructureComponentSummaryWithHistory">
+            <summary>
+            Summarizes a single infrastructure unit stats: score, trends, name.
+            </summary>
+        </member>
+        <member name="P:webapp.Models.InfrastructureComponentSummaryWithHistory.ScoreHistory">
             <summary>
             Holds Scores per last 30 days.
             If no data for a day - places 0.
             </summary>
         </member>
-        <member name="P:webapp.Models.InfrastructureComponentSummary.ScoreTrend">
+        <member name="P:webapp.Models.InfrastructureComponentSummaryWithHistory.ScoreTrend">
             <summary>
             Pre-calculated parameters for drawing trend line.
             </summary>
         </member>
-        <member name="M:webapp.Models.InfrastructureComponentSummary.#ctor">
+        <member name="T:webapp.Models.InfrastructureOverview">
             <summary>
-            Initializes a new instance of the <see cref="T:webapp.Models.InfrastructureComponentSummary"/> class.
-            Creates random id on constructor.
+            Returns infrastructure summary: scores by category, counters, etc.
+            </summary>
+        </member>
+        <member name="P:webapp.Models.InfrastructureOverview.Overall">
+            <summary>
+            Overall infrastructure summary.
+            </summary>
+        </member>
+        <member name="P:webapp.Models.InfrastructureOverview.Components">
+            <summary>
+            Separate summary for each involved component.
+            </summary>
+        </member>
+        <member name="T:webapp.Models.InfrastructureOverviewDiff">
+            <summary>
+            Returns infrastructure summary: scores by category, counters, etc.
+            </summary>
+        </member>
+        <member name="P:webapp.Models.InfrastructureOverviewDiff.Overall1">
+            <summary>
+            First overall infrastructure summary.
+            </summary>
+        </member>
+        <member name="P:webapp.Models.InfrastructureOverviewDiff.Overall2">
+            <summary>
+            Second overall infrastructure summary.
+            </summary>
+        </member>
+        <member name="P:webapp.Models.InfrastructureOverviewDiff.Components1">
+            <summary>
+            Components of first summary.
+            </summary>
+        </member>
+        <member name="P:webapp.Models.InfrastructureOverviewDiff.Components2">
+            <summary>
+            Components of second summary.
             </summary>
         </member>
         <member name="T:webapp.Models.ScoreHistoryItem">
@@ -264,46 +319,6 @@
             </summary>
             <param name="values">The score values.</param>
             <returns>Trend object.</returns>
-        </member>
-        <member name="T:webapp.Models.InfrastructureOverview">
-            <summary>
-            Returns infrastructure summary: scores by category, counters, etc.
-            </summary>
-        </member>
-        <member name="P:webapp.Models.InfrastructureOverview.Overall">
-            <summary>
-            Overall infrastructure summary.
-            </summary>
-        </member>
-        <member name="P:webapp.Models.InfrastructureOverview.Components">
-            <summary>
-            Separate summary for each involved component.
-            </summary>
-        </member>
-        <member name="T:webapp.Models.InfrastructureOverviewDiff">
-            <summary>
-            Returns infrastructure summary: scores by category, counters, etc.
-            </summary>
-        </member>
-        <member name="P:webapp.Models.InfrastructureOverviewDiff.Overall1">
-            <summary>
-            First overall infrastructure summary.
-            </summary>
-        </member>
-        <member name="P:webapp.Models.InfrastructureOverviewDiff.Overall2">
-            <summary>
-            Second overall infrastructure summary.
-            </summary>
-        </member>
-        <member name="P:webapp.Models.InfrastructureOverviewDiff.Components1">
-            <summary>
-            Components of first summary.
-            </summary>
-        </member>
-        <member name="P:webapp.Models.InfrastructureOverviewDiff.Components2">
-            <summary>
-            Components of second summary.
-            </summary>
         </member>
         <member name="T:webapp.Program">
             <summary>
