+++ src/frontend/src/services/DataService.ts
@@ -170,10 +170,10 @@ export class DataService {
   }
 
   public async getGeneralOverviewDiffData(date1: string, date2: string) {
-    console.log(`[] calling api/audits/overviewdiff/`);
+    console.log(`[] calling api/audits/overview/diff/`);
     let suffix = '?date1=' + encodeURIComponent(date1) + '&date2=' + encodeURIComponent(date2);
     return axios
-      .get(this.baseUrl + "/api/audits/overviewdiff" + suffix)
+      .get(this.baseUrl + "/api/audits/overview/diff" + suffix)
       .then((response) => response.data)
       .catch((error) => console.log(error))
       .finally(() => console.log("overview diff request finished."));
